move to package underneath pkg/network/usm
