+++ pkg/network/usm/tests/tracer_classification_linux_test.go
@@ -5,7 +5,7 @@
 
 //go:build linux_bpf
 
-package usm
+package tests
 
 import (
 	"testing"
