+++ pkg/network/usm/tests/tracer_usm_linux_test.go
@@ -5,7 +5,7 @@
 
 //go:build linux_bpf
 
-package usm
+package tests
 
 import (
 	"bytes"
